requirements.txt
pandas
streamlit
plotly
wordcloud
faiss-cpu
setuptools